+++ packages/gamut/src/List/List.tsx
@@ -1,10 +1,17 @@
 import { DotLoose } from '@codecademy/gamut-patterns';
+import { timingValues } from '@codecademy/gamut-styles';
 import isArray from 'lodash/isArray';
 import { ComponentProps, forwardRef, useEffect, useRef, useState } from 'react';
 import * as React from 'react';
 
 import { Box, BoxProps, FlexBox } from '../Box';
-import { ListEl, ListWrapper } from './elements';
+import {
+  AnimatedListWrapper,
+  hiddenVariant,
+  ListEl,
+  shadowVariant,
+  StaticListWrapper,
+} from './elements';
 import { ListProvider, useList } from './ListProvider';
 import { AllListProps } from './types';
 
@@ -26,11 +33,16 @@ export interface ListProps extends AllListProps<ComponentProps<typeof ListEl>> {
   /** A custom message to override the default empty message  */
   emptyMessage?: React.ReactNode;
   /**
-   * @deprecated Use overflow instead!
-   * Whether the List container should have overflow hidden.
+   * How the List container should handle overflow.
    */
-  overflowHidden?: boolean;
   overflow?: BoxProps['overflow'];
+  /**
+   * This is an override for the width of the wrapper element that contains the List.
+   * It is useful for custom scroll and breakpoint handling. Use with caution.
+   */
+  wrapperWidth?: BoxProps['width'];
+  /** Whether to disable container queries on the List wrapper */
+  disableContainerQuery?: boolean;
 }
 
 export const List = forwardRef<HTMLUListElement, ListProps>(
@@ -49,15 +61,22 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
       children,
       header,
       emptyMessage,
-      overflowHidden = false,
       overflow = 'auto',
       scrollToTopOnUpdate = false,
+      wrapperWidth,
+      disableContainerQuery = false,
+      ...rest
     },
     ref
   ) => {
     const isEmpty = !children || (isArray(children) && children.length === 0);
+    const isTable = as === 'table';
+
     const [isEnd, setIsEnd] = useState(false);
-    const showShadow = shadow && scrollable && !isEnd;
+    const ListWrapper = shadow ? AnimatedListWrapper : StaticListWrapper;
+    const showShadow = shadow && scrollable && !isEnd && !isEmpty;
+    const animationVar = showShadow ? 'shadow' : 'hidden';
+
     const value = useList({
       listType: as,
       rowBreakpoint,
@@ -77,7 +96,6 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
       setIsEnd(tableWidth < wrapperWidth);
     }, []);
 
-    const isTable = as === 'table';
     useEffect(() => {
       if (scrollToTopOnUpdate && tableRef.current !== null) {
         tableRef.current.scrollTo({ top: 0 });
@@ -85,7 +103,12 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
     });
 
     const listContent = (
-      <ListEl as={isTable ? 'tbody' : as} ref={ref} variant={value.variant}>
+      <ListEl
+        as={isTable ? 'tbody' : as}
+        ref={ref}
+        variant={value.variant}
+        {...rest}
+      >
         {children}
       </ListEl>
     );
@@ -124,12 +147,20 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
     return (
       <ListProvider value={value}>
         <ListWrapper
+          animate={animationVar}
+          disableContainerQuery={disableContainerQuery}
           id={id}
           maxHeight={height}
-          overflow={overflowHidden ? 'hidden' : overflow}
+          overflow={overflow}
           position="relative"
           ref={wrapperRef}
-          scrollable={!isEmpty && showShadow}
+          transition={{
+            background: { duration: timingValues.fast, ease: 'easeInOut' },
+          }}
+          variants={{
+            hidden: hiddenVariant,
+            shadow: shadowVariant,
+          }}
           width={1}
           onScroll={scrollable ? scrollHandler : undefined}
         >
@@ -137,19 +168,18 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
             as={isTable && !isEmpty && !loading ? 'table' : 'div'}
             data-testid={`scrollable-${id}`}
             height={isEmpty ? height : 'fit-content'}
-            maxWidth={1}
+            maxWidth={wrapperWidth || 1}
             minHeight={minHeight}
             overflow="inherit"
             position="relative"
             ref={!isEmpty ? tableRef : undefined}
-            width="inherit"
+            width={wrapperWidth || 'inherit'}
           >
             {content}
           </Box>
-
           {isEmpty && (
             <FlexBox center width={1}>
-              <DotLoose inset={0} position="absolute" top={-2} />
+              <DotLoose inset={0} position="absolute" top={0} />
             </FlexBox>
           )}
         </ListWrapper>