+++ packages/gamut/src/PopoverContainer/types.ts
@@ -83,4 +83,8 @@ export interface PopoverContainerProps
    * The target element around which the popover will be positioned.
    */
   targetRef: RefObject<TargetRef>;
+  /**
+   * If true, it will allow outside page interaction. Popover container will still close when clicking outside of the popover or hitting the escape key.
+   */
+  allowPageInteraction?: boolean;
 }