+++ packages/styleguide/.storybook/components/index.tsx
@@ -3,5 +3,6 @@ export * from './TokenTable';
 export * from './Headers';
 export * from './Elements/Callout';
 export * from './Elements/Markdown';
+export * from './Elements/ImageGallery';
 export * from './Scales/ColorScale';
 export * from './Navigation/TableOfContents';