+++ packages/styleguide/src/lib/Atoms/Animations/Rotation/Rotation.stories.tsx
@@ -1,7 +1,7 @@
 import { Rotation, StrokeButton } from '@codecademy/gamut';
 import { MiniChevronDownIcon } from '@codecademy/gamut-icons';
 import type { Meta } from '@storybook/react';
-import { useState } from 'react';
+import { ComponentProps, useEffect, useState } from 'react';
 
 const meta: Meta<typeof Rotation> = {
   component: Rotation,
@@ -10,12 +10,16 @@ const meta: Meta<typeof Rotation> = {
 
 export default meta;
 
-export const Default: React.FC = () => {
-  const [isRotated, setRotated] = useState(false);
+export const Default: React.FC<ComponentProps<typeof Rotation>> = (args) => {
+  const [isRotated, setRotated] = useState(args.rotated);
+
+  useEffect(() => {
+    setRotated(args.rotated);
+  }, [args.rotated]);
 
   return (
     <StrokeButton onClick={() => setRotated(!isRotated)}>
-      <Rotation rotated={isRotated}>
+      <Rotation {...args} rotated={isRotated}>
         <MiniChevronDownIcon />
       </Rotation>
     </StrokeButton>