+++ packages/styleguide/src/lib/Atoms/Icons/Regular/Regular.stories.tsx
@@ -3,6 +3,10 @@ import { GamutIconProps } from '@codecademy/gamut-icons';
 import * as icons from '@codecademy/gamut-icons/src/icons/regular';
 import type { Meta, StoryObj } from '@storybook/react';
 
+import { ImageGallery } from '~styleguide/blocks';
+
+import { LE_ICONS, SKILLS_ICONS, UI_ICONS, VENDOR_ICONS } from '../constants';
+
 type IconProps = GamutIconProps & {
   icon: typeof icons.AlertIcon;
 };
@@ -31,5 +35,29 @@ export const Default: Story = {
   args: {
     size: 40,
     icon: icons.AlertIcon,
+  },
+};
+
+export const RegularInterfaceIcons: Story = {
+  render: () => {
+    return <ImageGallery imageType="icon" images={UI_ICONS} />;
+  },
+};
+
+export const RegularLearningEnvironmentIcons: Story = {
+  render: () => {
+    return <ImageGallery imageType="icon" images={LE_ICONS} />;
+  },
+};
+
+export const RegularVendorIcons: Story = {
+  render: () => {
+    return <ImageGallery imageType="icon" images={VENDOR_ICONS} />;
+  },
+};
+
+export const RegularSkillIcons: Story = {
+  render: () => {
+    return <ImageGallery imageType="icon" images={SKILLS_ICONS} />;
   },
 };