+++ packages/styleguide/src/lib/Atoms/Illustrations/Illustrations.stories.tsx
@@ -2,6 +2,8 @@ import type { IllustrationProps } from '@codecademy/gamut-illustrations';
 import * as illustrations from '@codecademy/gamut-illustrations';
 import type { Meta, StoryObj } from '@storybook/react';
 
+import { ImageGallery } from '~styleguide/blocks';
+
 type IllustrationComponentProps = IllustrationProps & {
   illustration: React.ComponentType<IllustrationProps>;
 };
@@ -34,5 +36,23 @@ export const Default: Story = {
     width: 256,
     height: 256,
     illustration: illustrations.NumberBlocks,
+  },
+};
+
+export const AllIllustrations: Story = {
+  render: () => {
+    return (
+      <ImageGallery
+        controls={{
+          minColumns: 1,
+          columns: 4,
+          maxColumns: 5,
+          imageSize: 80,
+          maxImageSize: 500,
+        }}
+        imageType="illustration"
+        images={illustrations}
+      />
+    );
   },
 };