+++ packages/styleguide/src/lib/Organisms/ConnectedForm/ConnectedFormGroup/ConnectedFormGroup.stories.tsx
@@ -1,5 +1,7 @@
 import {
+  ConnectedForm,
   ConnectedFormGroup,
+  ConnectedFormGroupProps,
   ConnectedRadioGroupInput,
   useConnectedForm,
 } from '@codecademy/gamut';
@@ -8,13 +10,50 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 const meta: Meta<typeof ConnectedFormGroup> = {
   component: ConnectedFormGroup,
-  args: {},
+  args: {
+    field: {
+      component: ConnectedRadioGroupInput,
+      options: [
+        { label: 'one', value: 'one' },
+        { label: 'two', value: 'two' },
+        { label: 'zero', value: 'zero' },
+      ],
+    } as ConnectedFormGroupProps<typeof ConnectedRadioGroupInput>['field'],
+    label: 'i am a label',
+    name: 'radioGroupExample',
+  },
+  argTypes: {
+    loading: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
+  },
 };
 
 export default meta;
 type Story = StoryObj<typeof ConnectedFormGroup>;
 
-const FormGroupStates = () => {
+export const Default: Story = {
+  render: (args) => (
+    <ConnectedForm
+      mt={12}
+      resetOnSubmit
+      onSubmit={(values) => {
+        action('Form Submitted')(values);
+      }}
+    >
+      <ConnectedFormGroup {...args} />
+    </ConnectedForm>
+  ),
+};
+
+export const States = () => {
   const { ConnectedFormGroup, ConnectedForm, connectedFormProps } =
     useConnectedForm({
       defaultValues: {
@@ -79,8 +118,4 @@ const FormGroupStates = () => {
       />
     </ConnectedForm>
   );
-};
-
-export const Default: Story = {
-  render: () => <FormGroupStates />,
 };