+++ packages/styleguide/src/lib/Organisms/ConnectedForm/SubmitButton/SubmitButton.stories.tsx
@@ -5,12 +5,39 @@ import type { Meta, StoryObj } from '@storybook/react';
 const meta: Meta<typeof SubmitButton> = {
   component: SubmitButton,
   args: {},
+  argTypes: {
+    loading: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
+  },
 };
 
 export default meta;
 type Story = StoryObj<typeof SubmitButton>;
 
-const SubmitButtonStates = () => {
+export const Default: Story = {
+  render: (args) => (
+    <ConnectedForm
+      my={24}
+      onSubmit={(values) => {
+        action('Form Submitted')(values);
+      }}
+    >
+      <SubmitButton m={8} {...args}>
+        submit the form, please
+      </SubmitButton>
+    </ConnectedForm>
+  ),
+};
+
+export const States = () => {
   return (
     <ConnectedForm
       display="flex"
@@ -43,8 +70,4 @@ const SubmitButtonStates = () => {
       </FlexBox>
     </ConnectedForm>
   );
-};
-
-export const Default: Story = {
-  render: () => <SubmitButtonStates />,
 };