+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.50.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.49.2...@codecademy/gamut-icons@9.50.0) (2025-09-15)
+
+### Features
+
+- **Icon:** added new file-text-info icon. ([a06aeb4](https://github.com/Codecademy/gamut/commit/a06aeb44d6a9429eac7bc28d3483a690dbabf3bd))
+
+### [9.49.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.49.1...@codecademy/gamut-icons@9.49.2) (2025-08-20)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+### [9.49.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.49.0...@codecademy/gamut-icons@9.49.1) (2025-08-20)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+## [9.49.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.48.0...@codecademy/gamut-icons@9.49.0) (2025-08-12)
+
+### Features
+
+- **Icons:** add git, markdown, typescript icons ([4b15f3b](https://github.com/Codecademy/gamut/commit/4b15f3b5826f28c50ec782b3fb9e37e2a0bd797e))
+
+## [9.48.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.47.0...@codecademy/gamut-icons@9.48.0) (2025-08-08)
+
+### Features
+
+- Add icons for Apple, EA, Google, IBM, Instagram (rounder variant), Meta, Reddit (solid), Spotify and Windows ([0bdcd8f](https://github.com/Codecademy/gamut/commit/0bdcd8fe41d732414e2e186a1e4afb4eb5189f31))
+
+## [9.47.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.46.3...@codecademy/gamut-icons@9.47.0) (2025-08-07)
+
+### Features
+
+- **Icons:** Adds BookLibrary and HammerWrench icons ([36293b1](https://github.com/Codecademy/gamut/commit/36293b17f453314ea507fdec9ef428ce745d7546))
+
+### [9.46.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.46.2...@codecademy/gamut-icons@9.46.3) (2025-07-22)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+### [9.46.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.46.1...@codecademy/gamut-icons@9.46.2) (2025-07-17)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
 ### [9.46.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.46.0...@codecademy/gamut-icons@9.46.1) (2025-06-17)
 
 **Note:** Version bump only for package @codecademy/gamut-icons